replace redundant functions with generic ones

+++ src/problems/p_0024/mod.rs
@@ -16,7 +16,7 @@ fn solve() -> String {
     while permutations != 0 {
         // factorial of the number of digits minus 1
         // this is the number of permutations for digits after the first digit
-        let fact = factorial((digits.len() as u64) - 1);
+        let fact = factorial(digits.len() - 1);
 
         // we divide the number of permutations by the factorial
         // this is because if there are more permutations left than the factorial
@@ -25,13 +25,13 @@ fn solve() -> String {
         let index = permutations / fact;
 
         // add the digit to the result (since we are now taking it as the first digit)
-        result.push_str(&digits[index as usize].to_string());
+        result.push_str(&digits[index].to_string());
 
         // remove that digit from the list of digits
-        digits.remove(index as usize);
+        digits.remove(index);
 
         // because the initial vector is sorted, we don't need to sort digits again
-        // as  we aren't actually permutating the vector of digits
+        // as we aren't actually permutating the vector of digits
 
         // reduce the number of permutations left
         // note that if the index is zero, we aren't reducing the number of permutations
